requirements.txt
annotated-types==0.7.0
anyio==3.7.1
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.7
click==8.3.2
colorama==0.4.6
cryptography==46.0.7
dotenv==0.9.9
fastapi==0.104.1
google-api-core==2.30.2
google-auth==2.49.1
google-cloud-core==2.5.1
google-cloud-storage==2.13.0
google-crc32c==1.8.0
google-resumable-media==2.8.2
googleapis-common-protos==1.74.0
greenlet==3.1.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.11
jsonpatch==1.33
jsonpointer==3.1.1
langchain-core==1.2.28
langgraph==1.1.6
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.9
langgraph-sdk==0.3.13
langsmith==0.7.30
orjson==3.11.8
ormsgpack==1.12.2
packaging==23.2
playwright==1.58.0
proto-plus==1.27.2
protobuf==7.34.1
pyasn1==0.6.3
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
pyee==13.0.1
python-dotenv==1.2.2
python-multipart==0.0.6
PyYAML==6.0.3
requests==2.31.0
requests-toolbelt==1.0.0
sniffio==1.3.1
starlette==0.27.0
tenacity==8.5.0
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.3
uuid_utils==0.14.1
uvicorn==0.24.0.post1
xxhash==3.6.0
zstandard==0.25.0
